feat: ✨ Add lock initiative button
After some consideration, I believe the cleanest way to allow users to use functionality like the 'Veteran'-ability in Dragonbane is to implement it in Year Zero Combat itself (it requires changing how some internal methods execute and we cannot expose that in a configurable manner).

The buttons for locking initiative will now appear by default when 'Reset the Initiative Each Round' is selected.

The resetInitiativeDeck has seen some change in it's functionality as a result. Instead of using 'recall' we now use 'pass' and supply a filtered list of cards this allows us to keep the initiative cards we want to keep between draws in the discard pile.